--- ibl-frontend/frontend-content/src/app/session-list/GithubApiInterface.ts
@@ -1,6 +1,0 @@
-import { GithubIssue } from "./GithubIssueInterface";
-
-export interface GithubApi {
-    records: GithubIssue[];
-    records_count: number;
-  }
