--- ibl-frontend/frontend-content/src/app/session-list/GithubIssueInterface.ts
@@ -1,20 +1,0 @@
-export interface GithubIssue {
-    mouse_id: number;
-    session_date: string;
-    session_lab: string;
-    subject_nickname: string;
-    subject_birth_date: string;
-    session_start_time: string;
-    task_protocol: string; 
-    subject_line: string; 
-    responsible_user: string;
-    session_uuid: string; 
-    sex: string;
-    subject_uuid: string;
-    nplot: string;
-    nprobe: string;
-    session_project: string;
-    good4bmap: string;
-  }
-
-
