+++ ibl-frontend/frontend-content/src/app/session-list/all-sessions.service.ts
@@ -1,13 +1,34 @@
 import { Injectable } from '@angular/core';
 import { Subject, Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
-import { GithubApi } from './GithubApiInterface';
-import { GithubIssue } from './GithubIssueInterface';
-
 import { environment } from '../../environments/environment';
 
 
 const BACKEND_API_URL = environment.backend_url;
+
+export interface GithubIssue {
+  mouse_id: number;
+  session_date: string;
+  session_lab: string;
+  subject_nickname: string;
+  subject_birth_date: string;
+  session_start_time: string;
+  task_protocol: string; 
+  subject_line: string; 
+  responsible_user: string;
+  session_uuid: string; 
+  sex: string;
+  subject_uuid: string;
+  nplot: string;
+  nprobe: string;
+  session_project: string;
+  good4bmap: string;
+}
+
+export interface GithubApi {
+  records: GithubIssue[];
+  records_count: number;
+}
 
 @Injectable({
   providedIn: 'root'
